chore: rename Settings service to Config service
