+++ src/contexts/ServicesContext.jsx
@@ -4,7 +4,7 @@ import {
   PageService,
   DirectoryService,
   MoverService,
-  SettingsService,
+  ConfigService,
   MediaService,
 } from "services"
 
@@ -14,14 +14,14 @@ const ServicesProvider = ({ client, children }) => {
   const pageService = new PageService({ apiClient: client })
   const directoryService = new DirectoryService({ apiClient: client })
   const moverService = new MoverService({ apiClient: client })
-  const settingsService = new SettingsService({ apiClient: client })
+  const configService = new ConfigService({ apiClient: client })
   const mediaService = new MediaService({ apiClient: client })
 
   const services = {
     pageService,
     directoryService,
     moverService,
-    settingsService,
+    configService,
     mediaService,
   }
