+++ src/services/index.js
@@ -2,4 +2,4 @@ export { PageService } from "./PageService"
 export { MediaService } from "./MediaService"
 export { DirectoryService } from "./DirectoryService"
 export { MoverService } from "./MoverService"
-export { SettingsService } from "./SettingsService"
+export { ConfigService } from "./ConfigService"
